pyproject.toml
[tool.poetry]
name = "inventory"
version = "0.1.0"
description = "Inventory management system"
authors = ["Thom Wiggers <thom@thomwiggers.nl>"]
license = "AGPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
python-stdnum = "^1.11"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
